first draft of re-implementing door using api gateway and asf auth api
