+++ buildspec.yml
@@ -1,30 +1,13 @@
 version: 0.2
 
-env:
-  variables:
-    PIPELINE_NAME: "distribution-test"
-    IMAGE_NAME: "door"
-    IMAGE_HOST: "406893895021.dkr.ecr.us-east-1.amazonaws.com"
-    MATURITY: "TEST"
-
 phases:
-  pre_build:
-    commands:
-      - $(aws ecr get-login --region $AWS_DEFAULT_REGION --no-include-email)
-      - aws codepipeline get-pipeline-state --name $PIPELINE_NAME --query 'stageStates[*].actionStates[?actionName==`Source`].currentRevision.revisionId' --output text | cut -c1-7 > sha1.txt
-      - sed -i "s/\$SHA1/$(cat sha1.txt)/" cloudformation.yaml
+  install:
+    runtime-versions:
+      python: 3.9
   build:
     commands:
-      - docker build -t $IMAGE_NAME door/
-      - docker tag $IMAGE_NAME:latest $IMAGE_HOST/$IMAGE_NAME:latest
-      - docker tag $IMAGE_NAME:latest $IMAGE_HOST/$IMAGE_NAME:$MATURITY
-      - docker tag $IMAGE_NAME:latest $IMAGE_HOST/$IMAGE_NAME:$(cat sha1.txt)
+      - python -m pip install -r door/requirements.txt -t door/src/
       - aws cloudformation package --template-file cloudformation.yaml --s3-bucket $BUCKET --s3-prefix $PREFIX --output-template-file cloudformation-final.yaml
-  post_build:
-    commands:
-      - docker push $IMAGE_HOST/$IMAGE_NAME:latest
-      - docker push $IMAGE_HOST/$IMAGE_NAME:$MATURITY
-      - docker push $IMAGE_HOST/$IMAGE_NAME:$(cat sha1.txt)
 artifacts:
   type: zip
   files:
