+++ cloudformation.yaml
@@ -2,14 +2,6 @@ AWSTemplateFormatVersion: 2010-09-09
 
 Parameters:
 
-  VpcId:
-    Description: VPC Id in which to launch EC2 instances for ingest activities
-    Type: AWS::EC2::VPC::Id
-
-  SubnetId:
-    Description: List of Subnet Ids in which to launch EC2 instances for ingest activities
-    Type: List<AWS::EC2::Subnet::Id>
-
   PrivateBucket:
     Description: Name of S3 bucket in which zipped output products will be archived
     Type: String
@@ -17,23 +9,10 @@ Parameters:
   LogBucket:
     Type: String
 
-  DoorContainerRepo:
-    Description: URL for docker container repository for door application (repository-url/image)
-    Type: String
-
-  UrsClientId:
-    Type: String
-
-  UrsAuthCode:
-    Type: String
-
   CertificateArn:
     Type: String
 
-  LoadBalancerCidrIp:
-    Type: String
-
-  Hostname:
+  DomainName:
     Type: String
 
   CloudFrontKeyPairId:
@@ -42,16 +21,14 @@ Parameters:
   PrivateKeySecretName:
     Type: String
 
-Outputs:
+  AuthUrl:
+    Type: String
 
-  AppUrl:
-    Value: !GetAtt DoorStack.Outputs.AppUrl
+  JwtCookieName:
+    Type: String
 
-  LoadBalancerDnsName:
-    Value: !GetAtt DoorStack.Outputs.LoadBalancerDnsName
-
-  UrsRedirectUri:
-    Value: !GetAtt DoorStack.Outputs.UrsRedirectUri
+  JwtPublicKey:
+    Type: String
 
 Resources:
 
@@ -60,18 +37,15 @@ Resources:
     Properties:
       Parameters:
         Name: !Sub "${AWS::StackName}-door"
-        VpcId: !Ref VpcId
-        SubnetId: !Join [",", !Ref SubnetId]
         CloudFrontDomainName: !GetAtt CloudFrontStack.Outputs.DomainName
         CloudFrontKeyPairId: !Ref CloudFrontKeyPairId
         PrivateKeySecretName: !Ref PrivateKeySecretName
-        DoorContainerImage: !Sub "${DoorContainerRepo}:$SHA1"
         PrivateBucket: !Ref PrivateBucket
-        UrsClientId: !Ref UrsClientId
-        UrsAuthCode: !Ref UrsAuthCode
         CertificateArn: !Ref CertificateArn
-        LoadBalancerCidrIp: !Ref LoadBalancerCidrIp
-        Hostname: !Ref Hostname
+        DomainName: !Ref DomainName
+        AuthUrl: !Ref AuthUrl
+        JwtCookieName: !Ref JwtCookieName
+        JwtPublicKey: !Ref JwtPublicKey
       TemplateURL: door/cloudformation.yaml
 
   CloudFrontStack:
