--- door/conf/apache2.conf
@@ -1,63 +1,0 @@
-Mutex file:${APACHE_LOCK_DIR} default
-PidFile ${APACHE_PID_FILE}
-
-User ${APACHE_RUN_USER}
-Group ${APACHE_RUN_GROUP}
-
-Timeout 300
-KeepAlive Off
-
-HostnameLookups Off
-
-IncludeOptional mods-enabled/*.load
-IncludeOptional mods-enabled/*.conf
-
-Listen 80
-
-DocumentRoot "/var/www/html"
-
-<Location "/">
-  Require all granted
-</Location>
-
-BrowserMatchNoCase "mozilla" BROWSER_USER_AGENT
-
-<IfDefine !NO_AUTH>
-  ServerName ${SERVER_NAME}
-  LoadModule auth_urs_module /usr/lib/apache2/modules/mod_auth_urs.so
-  UrsSessionStorePath /var/tmp/urs/session
-  UrsAuthServer ${URS_AUTH_SERVER}
-  UrsAuthPath /oauth/authorize
-  UrsTokenPath /oauth/token
-
-  <Location "/door">
-    AuthType UrsOAuth2
-    Require valid-user
-    UrsAuthGroup GRFNDoor
-    UrsClientId ${URS_CLIENT_ID}
-    UrsAuthCode ${URS_AUTH_CODE}
-    UrsRedirectUrl ${URS_REDIRECT_URL}
-    UrsIPCheckOctets 2
-    UrsUserProfileEnv uid URS_USERID
-    UrsUserProfileEnv email_address URS_EMAIL
-    <If "-z reqenv('BROWSER_USER_AGENT')">
-      Urs401Enable true
-    </If>
-  </Location>
-</IfDefine>
-<IfDefine NO_AUTH>
-  PassEnv URS_USERID
-  PassEnv URS_EMAIL
-</IfDefine>
-
-RedirectMatch 301 ^/$ /door/
-
-LogFormat "%h %l %u %t \"%r\" %>s %O \"%{Referer}i\" \"%{User-Agent}i\"" combined
-ErrorLog /dev/stderr
-CustomLog /dev/stdout combined
-LogLevel warn
-
-PassEnv DOOR_CONFIG
-
-WSGIDaemonProcess door group=${APACHE_RUN_GROUP} threads=5 user=${APACHE_RUN_USER}
-WSGIScriptAlias /door "/var/www/door/door.wsgi"
