--- door/conf/docker_run_with_urs.env
@@ -1,14 +1,0 @@
-SERVER_NAME=grfn-door-test.asf.alaska.edu
-URS_AUTH_SERVER=https://urs.earthdata.nasa.gov
-URS_CLIENT_ID=hmKbZuUjez9JyjTg389A5A
-URS_AUTH_CODE=change_me
-URS_REDIRECT_URL=https://grfn-door-test.asf.alaska.edu/oauth
-DOOR_CONFIG='{
-  "aws_region": "us-east-1",
-  "bucket_name": "gsfc-ngap-asf-grfn-private-test",
-  "expire_time_in_seconds": 15,
-  "users_table": "distribution-test-users",
-  "availability_lambda": "distribution-test-glacier-availability",
-  "status_lambda": "distribution-test-glacier-status",
-  "retention_days": 7
-}'
