--- door/conf/flask_run.env
@@ -1,12 +1,0 @@
-FLASK_APP=door.py
-URS_USERID=asjohnston
-URS_EMAIL=asjohnston@alaska.edu
-DOOR_CONFIG='{
-  "aws_region": "us-east-1",
-  "bucket_name": "gsfc-ngap-asf-grfn-private-test",
-  "expire_time_in_seconds": 15,
-  "users_table": "distribution-test-users",
-  "availability_lambda": "distribution-test-glacier-availability",
-  "status_lambda": "distribution-test-glacier-status",
-  "retention_days": 7
-}'
