+++ door/requirements.txt
@@ -1,3 +1,4 @@
-boto3==1.12.3
-Flask==1.1.1
-rsa==4.0
+boto3
+Flask
+rsa
+serverless_wsgi
