+++ door/src/door/__main__.py
@@ -1,0 +1,4 @@
+from door import app
+
+if __name__ == "__main__":
+    app.run(port=8080)
