--- door/src/html/healthcheck.html
@@ -1,8 +1,0 @@
-<html>
-<head>
-</head>
-<body>
-Apache is running
-</body>
-</html>
-
